linkedListCollection: refactor: create LinkedListOperation new file
